renamed SVMTest to main
